Refactor to rename `common.ts` to `helpers.ts`
